Extract `ShowController` from `DecisionsController`
We do this rather than further growing `DecisionsController`. `DecisionsController` now relates to viewing and downloading a collection of datasets, while `ShowController` will related to viewing and downloading a single dataset
